+++ web/static/js/containers/ShowIdeaContainer.jsx
@@ -1,15 +1,23 @@
-import React, { PropTypes } from 'react';
+import React, { PropTypes, Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import ShowIdea from '../components/ShowIdea';
 import { fetchIdea, addComment } from '../actions/ideas';
 
-const ShowIdeaContainer = ({idea, fetchIdea, onAddComment, ideaId, currentUser}) => {
-  if (!idea || idea.id != ideaId) {
+class ShowIdeaContainer extends Component {
+  componentWillMount() {
+    const { fetchIdea, ideaId } = this.props;
     fetchIdea(ideaId);
-    return null;
-  } else {
-    return (<ShowIdea idea={idea} currentUser={currentUser} onAddComment={(body) => onAddComment(idea.id, body)}/>);
+  }
+
+  render () {
+    const {idea, onAddComment, currentUser} = this.props;
+
+    if (idea === null || idea.loading) {
+      return null
+    } else {
+      return (<ShowIdea idea={idea} currentUser={currentUser} onAddComment={(body) => onAddComment(idea.id, body)}/>);
+    }
   }
 }