+++ web/static/js/reducers/currentIdea.js
@@ -1,9 +1,11 @@
-import { FETCH_IDEA_SUCCESS } from '../actions/ideas';
+import { FETCH_IDEA_REQUEST, FETCH_IDEA_SUCCESS } from '../actions/ideas';
 
 const currentIdea = (state = null, action) =>{
   switch(action.type) {
+    case FETCH_IDEA_REQUEST:
+      return { loading: true };
     case FETCH_IDEA_SUCCESS:
-      return { ...action.idea };
+      return { loading: false, ...action.idea };
     default:
       return state
   }