+++ web/static/js/actions/ideas.js
@@ -41,6 +41,7 @@ export const NEW_IDEA_RECEIVED = 'NEW_IDEA_RECEIVED';
 export const EDIT_IDEA_RECEIVED = 'EDIT_IDEA_RECEIVED';
 export const CHANGE_MY_VOTE = 'CHANGE_MY_VOTE';
 export const CHANGE_FILTER_TEXT = 'CHANGE_FILTER_TEXT';
+export const CHANGE_FILTER_STATUS = 'CHANGE_FILTER_STATUS';
 
 export const ALL_USERS_RECEIVED = 'ALL_USERS_RECEIVED';
 
@@ -84,6 +85,7 @@ const changeMyVote = (ideaId, myVote) => ({ type: CHANGE_MY_VOTE, ideaId, myVote
 const allUsersReceived = (users) => ({ type: ALL_USERS_RECEIVED, users });
 
 export const changeFilterText = (filter) => ({ type: CHANGE_FILTER_TEXT, filter });
+export const changeFilterStatus = (status) => ({ type: CHANGE_FILTER_STATUS, status});
 
 export const getAllUsers = () => (
   dispatch => {